EC/ROCM: fix clang rocm5.6 errors (#804)
this commit fixes two issues exposed by the UCC Linter-ROCM script.
First, there was a mistake on the configure test checking for the rocm
version. In addition, the hip_fp16 header file generates some warnings
with clang-12 that are not ignored with the -Werror flag being set.
Silence that particular warning for clang.