TOPO: Node ldr ordered by team (#1129)
## What

This PR sorts the node leader subgroup rank mapping by team rank to
avoid unnecessary unpacking in allgatherv. It also removes the
`is_host_ordered` function in the allgatherv code, because it was
previously checking whether or not the leaders were sorted by team rank.
Now it always will be.

## Why ?

Previously the node leader subgroup was sorted by hostid, meaning that
if a hostname happened to be lower than another hostname by the
comparator that UCC uses, the node leader allgatherv step in my cl/hier
allgatherv might actually be incorrect and need the extra unpack step to
rearrange the buffer to have the right order.

Previous PR for reference here: https://github.com/openucx/ucc/pull/1103

## How ?
Sorts the rank mapping and updates group_rank. In addition to removing
`is_host_ordered`, I also updated `is_block_ordered`, which previously
checked if the FULL_ORDERED subgroup was the same as the current team.
This was a tighter restriction than it needed to be, since FULL_ORDERED
also sorts by the socket, and the allgatherv only cares whether ranks
are block ordered on each node.

## Data


https://docs.google.com/spreadsheets/d/1icnHUZFCYvThZ_bsZ3PS3fCL5HLG58fGrpViF25dHDY/edit?usp=sharing

---------

Co-authored-by: Nicholas Sarkauskas <nsarkauskas@nsarkauskas-mlt.client.nvidia.com>