EC/CUDA: Binary search max kernel threads (#1166)
In UCC we use a kernel to handle copies and reductions for cuda buffers.
The kernel is failing to launch because it requires Registers Per Thread: 156
according to an nsys profile. By default we launch with 512 threads, so
512 * 156 = 79872, greater than the 64k limit.

So, the this PR will binary search to find the maximum threads per block that will fit the kernel.